Set PC WData correctly on xRET/exc
When an xRET or exception is encountered, the controller goes into the
FLUSH state. RVFI data is read while the controller is in the FLUSH
state, but the PC is only available the next cycle, so for RVFI we
need to set the PC WData again once it is available.