Set MTVEC to 0 on boot
In typical uses, MTVEC can be set after the core fetches the first
instruction and we can use the boot_addr_i signal to set it.
However, when using TestRIG it is possible that the first
instruction causes an exception, and in this case there is not enough
time to set MTVEC before the exception is handled, so the exception
vector will be whatever was in boot_addr_i on boot.
This is incompatible with TestRIG, so MTVEC is set to 0 on boot rather
than using boot_addr_i.