[rtl] Add SCR interface and initial DDC impl
This commit adds the SCR interface and connects it between ID and the CS
register file.
This commit also adds an initial implementation of DDC as a SCR. It is
writeable via the SCR interface. DDC is also muxed in as either operand
A or operand B in certain instructions; this commit adds that muxing.

Change-Id: I76b7fc2491c9a2130aa3385ccc05d8587c67737c