[rtl] Add unused PCC signals
This commit adds signals for the PCC in several locations; at the moment
this is unused. It is generated by setting the PC as the offset of an
almighty capability in the IF stage, and then propagated the same way
that the PC is propagated. Eventually logic will move to primarily use
the PCC rather than the PC; at this point, PC will be generated by
taking the offset of PCC.

Change-Id: I787ed9b9925edb0ccafe1ce3727800e86585e3d4