[rtl] Improve PCC functionality
There is now an explicit pcc_q signal which is used to store the PCC.
This is used mostly for the metadata. The IF stage submodules (cache or
prefetch buffer) keep track of the PC internally, so the IF stage
continuously updates the PCC so that its offset tracks the PC that the
submodules report.

Change-Id: I66c0eb9a7fc16f572bd1cf09418b54b66df16911