[rtl] Implement CInvoke and fix CJALR
One important thing is that previously jumps could not throw exceptions;
now they can, so changes are needed in the IF and ID stages to handle
this. In IF, we need to make sure we don't set the PC on jumps if it
encountered an exception. In ID, we need to make sure we go to
FIRST_CYCLE when a multicycle instruction gets an exception.

Note: PCC's bounds are still unused
Change-Id: I7c40b293dbca9f686bf0442ebeb4e13ad6ae6ede