[rtl] Add CHERI memory access exceptions
This commit adds a CHERI memory checker that takes a capability as an
input and checks that the request going through it is allowed.
Previously, all exceptions in the LSU required a memory response; this
is no longer true since misaligned capability accesses do not make it
out to the bus.

Note: At the moment, accesses that cause CHERI exceptions (not
misaligned) still go out on the bus. This will be fixed
in a later commit.

Change-Id: I5915a91ce5a44f72ec31a0f97169959d58320b78