[rtl] Fix CHERI instruction fetch exception
Only check the "upper" exception signal if this was a full instruction
fetch that started on a 4-byte-aligned boundary (if it was a compressed
fetch then we only need the bottom half, and if it was 4-byte-unaligned
then we only need the bottom half because the rest of the instruction
was fetched in the previous aligned read).
Note: this has not been tested with compressed instruction fetching,
only with unaligned full instruction fetches.

Change-Id: I38183f342fa02f71c14015b7e3b7fdcc02eb2dae