[rtl] Properly handle PC being PCC offset in IF
With CHERI, PC is defined as being the _offset_ of PCC. However, the
modules in the IF stage that handle actually fetching instructions and
setting the memory interface signals work with addresses. This cleans up
the variable names to more explicitly state whether the variable is an
offset (equivalent to a PC) or an address. It also adds logic that
correctly converts between the two.
This approach requires the fewest modifications to the internal logic of
the IF stage but might be wasteful - it might be possible to improve the
logic if the modules use offsets rather than addresses, but this change
would require additional care (for example to ensure that addresses
requested on the interface are still aligned when PCC's base is not
4-byte aligned.

Change-Id: If965b3d3c872ae3244307f71c9c177c44addefe7