[rtl] Fix instruction fetch CHERI checking logic
Edge cases could occur where you would fetch the upper half of a word
with a capability that only allowed you to fetch the lower half. Also,
with DII, the second half of a misaligned fetch is skipped, so this adds
logic that checks whether the second half would have been allowed.

Change-Id: I73597240c8ebe2ac4a422a1074919905c278861e