[rtl] Save IFetch CHERI errors in mcause + mtval
This adds logic in the Instruction Fetch stage to track whether there is
an error in the fetch FIFO, and whether the error is a CHERI error or
not. When it is, the CHERI exception information is passed to the
controller.
Since errors will cause FIFO flushes when they reach the controller,
any later bus errors (which this logic does not handle) will be flushed
and will not happen, so only one error needs to be tracked at a time.
A new signal is connected from the fetch FIFO to the IF stage to signal
when the output of the FIFO is the same as its input -- in this case the
exception information comes directly from the input to the IF stage
rather than from the registered exception signals.

Change-Id: I2cf40bbaf62c6fe547dc9ae71cb7744eaf8881a7