[rtl] Split PCC into two in IF stage
This commit splits the Program Counter Capability in the Instruction
Fetch stage into:
 - a jump PCC (which assumes a jump will be taken); and
 - a nojump PCC (which assumes a jump is not taken). This is also
   the PCC of the instruction currently being returned by the
   prefetcher/caches

The new_pcc muxes between the jump and nojump PCCs depending on whether
there has been a jump or not.

The new_pcc is used to provide authority for the current instruction
fetch (since it needs to account for the new bounds) and the nojump PCC
is used the same way as the old pc_if_o signal, so that it can be saved
to MEPC/MEPCC when required.

Change-Id: I7939efde4804872dc1b50eadd451c835fc399692