[rtl] Use address of new PCC for new instr address
This improves the critical path.
The previous implementation used the base of the new PCC, which requires
significant logic, and then added this to the new fetch offset, despite
the fact that the old logic was already factoring in the new offset into
the new PCC. Getting the address of the new PCC (which has the new fetch
offset factored in already) requires less logic.

Change-Id: Iad2b14371c3d1c7b4b7d5fbb3e94c20050f54fc3