[rtl] Use length in exception checking on JAL/JALR
The value calculated by the ALU uses the PC, which is an offset of the
current PCC. This means that we should compare it to the length rather
than the base and top.
Note: the comment explains that cheri-cap-lib does not rule out the
possibility of the base being above the top, and also does not explain
what happens to the length in this case.

Change-Id: I92fffcee60671f53addd64f0e79f17efe7f04a68