[rtl] Overhaul CHERI memory checking
This makes the memchecker modules purely combinational, and changes the
data and instruction interfaces to not issue requests on the bus that
will cause a CHERI exception.
On the instruction interface, requests that might cause an exception are
still issued because it is not known when the request is being made
whether the request is allowed or not (because we don't know whether the
resulting data contains a compressed instruction or not, and we don't
know until the instruction leaves the fetch FIFO which half of the
4-byte-aligned fetch it came from).

The "possible" exception signals are propagated through the fetch FIFO
with the data.

Fetches that are guaranteed to cause a CHERI exception (ones where both
halves of the fetch are disallowed) halt the prefetcher and prevent it
from making any further requests. Instructions in the fetch FIFO are
executed and if none of these flush the FIFO then the error will be
handled, which will flush the FIFO and restart prefetching.

Change-Id: I50c9c65d55c69e3131722708799eb330bada6f2b