[rtl] Don't check for IFetch exceptions before req
This change always issues requests on the instruction fetch interface
regardless of whether they cause a CHERI exception or not. The exception
signals are then passed through the prefetch buffer and fetch fifo the
same way they were passed through before (but as a struct now).
This improves the critical path. (~ -3.8ns slack before to ~ -2.73ns).

Change-Id: I9e68f4f47da730afb9d152805f2147fa50db6a06