[rtl] Remove cheri_alu->IF paths on CHERI jumps
Previously the path of CHERI jumps would go through the ALU, perform a
setAddress, then go to the IF stage, perform a getOffset, perform a
setOffset and then go through the memory checking.
Now, the target capability goes directly from ID to IF without going
through the CHERI ALU, and the IF stage sets the new offset and unseals
the capability. The same exception checking is still performed in the
CHERI ALU, and the CHERI ALU still needs to set the integer ALU inputs
so that the new offset can be calculated and passed to the IF stage.

Change-Id: I93983494f7f47a8481693cdbdc6fa2f673dc8fa3