[rtl] Add TestRIG compile time parameter
The changes in the IF stage for boot address and how to change the
address to be fetched on compressed instructions are only needed when
compiling for TestRIG, so this commit (finally) makes this a compile
time parameter.
The parameter is of "int" type because as far as I can tell, it's not
possible to have it set to 1 if it's a bit. FuseSoC uses the verilator
edalize backend, which passes the parameter as a literal integer using
verilator's -G argument
(https://github.com/olofk/edalize/blob/4044f2de5bf7f43e018e18f881468baebb3452c2/edalize/verilator.py#L165)
without giving it a size. Verilator interprets literal integer arguments
as 32bit signed integers
(https://github.com/verilator/verilator/blob/038d57070b539b2da2658edd7740d41ea069fd4f/src/V3AstNodes.cpp#L311)
so it sticks "32'h" before the integer when passing the argument in to a
module, and then complains because it is trying to set a 1 bit value to a
32 bit value.

Change-Id: If2a147f52df98f5bdedf351279a45cc96f0dfbb3